requirements.txt
numpy==1.18.4
opencv-python==4.2.0.34
Pillow==7.1.2
tflite-runtime @ https://dl.google.com/coral/python/tflite_runtime-2.1.0.post1-cp37-cp37m-linux_x86_64.whl
